participant-state: Move metrics-related code to another Bazel package.
