+++ ledger/participant-state-metrics/src/main/scala/com/daml/ledger/participant/state/metrics/Metrics.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.v1
+package com.daml.ledger.participant.state.metrics
 
 import java.util.concurrent.CompletionStage
