+++ ledger/participant-state-metrics/src/main/scala/com/daml/ledger/participant/state/v1/metrics/TimedReadService.scala
@@ -1,11 +1,13 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.v1
+package com.daml.ledger.participant.state.v1.metrics
 
 import akka.NotUsed
 import akka.stream.scaladsl.Source
 import com.codahale.metrics.MetricRegistry
+import com.daml.ledger.participant.state.metrics.Metrics
+import com.daml.ledger.participant.state.v1.{LedgerInitialConditions, Offset, ReadService, Update}
 import com.digitalasset.ledger.api.health.HealthStatus
 
 final class TimedReadService(delegate: ReadService, metrics: MetricRegistry, prefix: String)
