+++ ledger/participant-state/kvutils/app/BUILD.bazel
@@ -33,6 +33,7 @@ da_scala_library(
         "//ledger/ledger-api-health",
         "//ledger/participant-state",
         "//ledger/participant-state-index",
+        "//ledger/participant-state-metrics",
         "//ledger/participant-state/kvutils",
         "//ledger/sandbox",
         "//libs-scala/contextualized-logging",
